fix: avoid alloc on fill to capacity
We resize after a Push if the queue buffer is full.
This ensures that head and tail only meet if the queue is empty
so there's no need to otherwise differentiate between empty and full.

This introduces a small annoyance, though:
if one uses `NewQ(5)` and adds 5 items to the queue,
they expect no allocations,
but the current implementation will allocate.

Resolve this by always allocating N+1 slots
so that a full queue doesn't resize.